Day code phan cham cong va tinh Luong

+++ BaketyManagement/View/Forms/FrmTimeKeeping.cs
@@ -6,6 +6,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
 using System.Drawing;
+using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -34,11 +35,11 @@ namespace BaketyManagement.View.Forms
 
         private void LoadSalary()
         {
-            dgvTimeKeeping.Rows.Clear();
+
             row = 0;
             var query = from slr in db.Salaries
                         select new
-                        { 
+                        {
                             slr.IdStaff,
                             nameStaff = (from stf in db.staff
                                          where stf.IdStaff == slr.IdStaff
@@ -50,6 +51,7 @@ namespace BaketyManagement.View.Forms
                             slr.Rewards,
                             slr.SalaryOverTime,
                         };
+            dgvTimeKeeping.Rows.Clear();
             foreach (var slr in query)
             {
                 DateTime timekeep = (DateTime)slr.TimeKeeped;
@@ -91,7 +93,47 @@ namespace BaketyManagement.View.Forms
 
         private void btnTimeKeeping_Click(object sender, EventArgs e)
         {
-            
+            timeKeeping();
+        }
+
+        private void timeKeeping()
+        {
+            try
+            {
+                if (row < 0 || row >= dgvTimeKeeping.Rows.Count - 1)
+                {
+                    throw new Exception("Vui lòng chọn nhân viên cần cập nhật!");
+                }
+                Int32 idStaff = Convert.ToInt32(dgvTimeKeeping.Rows[row].Cells[0].Value);
+
+                var query = from slr in db.Salaries
+                            where slr.IdStaff == idStaff
+                            select slr;
+                string nowDay = DateTime.Now.ToString("dd/MM/yyyy");
+                foreach (var day in query)
+                {
+                    DateTime timeKeep = (DateTime)day.TimeKeeped;
+                    string timeKeepDay = timeKeep.ToString("dd/MM/yyyy");
+                    if (timeKeepDay == nowDay)
+                    {
+                        throw new Exception("Nhân viên này đã chấm công ngày hôm nay");
+                    }
+                }
+
+                Salary sl = (from slry in db.Salaries
+                             where slry.IdStaff == idStaff
+                             select slry).FirstOrDefault();
+                int m = (int)sl.WorkDay;
+                sl.WorkDay = m + 1;
+                sl.TimeKeeped = DateTime.Now;
+                db.SaveChanges();
+                MessageBox.Show("Chấm công thành công", "", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                LoadSalary();
+            }
+            catch (Exception ex)
+            {
+                MessageBox.Show(ex.Message);
+            }
         }
 
         private void editSalary()
@@ -127,15 +169,31 @@ namespace BaketyManagement.View.Forms
 
         private void btnCancelTimeKeeping_Click(object sender, EventArgs e)
         {
+            DialogResult result = MessageBox.Show("Bạn thực sự muốn xóa chấm công của nhân viên " + dgvTimeKeeping.Rows[row].Cells[1].Value.ToString(), "", MessageBoxButtons.YesNo, MessageBoxIcon.Question);
+            if (result == DialogResult.Yes)
+            {
+                Int32 idStaff = Convert.ToInt32(dgvTimeKeeping.Rows[row].Cells[0].Value);
+                Salary slXoa = (from sl in db.Salaries
+                                where sl.IdStaff == idStaff
+                                select sl).FirstOrDefault();
+                db.Salaries.Remove(slXoa);
+                db.SaveChanges();
+                row--;
+                MessageBox.Show("Xóa chấm công thành công", "", MessageBoxButtons.OK, MessageBoxIcon.Information);
+                LoadSalary();
+            }
+        }
+
+        private void btnFilter_Click(object sender, EventArgs e)
+        {
+            CultureInfo cul = CultureInfo.GetCultureInfo("vi-VN");
             Int32 idStaff = Convert.ToInt32(dgvTimeKeeping.Rows[row].Cells[0].Value);
-            Salary slXoa = (from sl in db.Salaries
+            Salary slTinh = (from sl in db.Salaries
                             where sl.IdStaff == idStaff
                             select sl).FirstOrDefault();
-            db.Salaries.Remove(slXoa);
-            db.SaveChanges();
-            row--;
-            MessageBox.Show("Xóa chấm công thành công", "", MessageBoxButtons.OK, MessageBoxIcon.Information);
-            LoadSalary();
+            Double tong = (double)((slTinh.WorkDay * slTinh.SalaryDate) + (slTinh.HoursOverTime * slTinh.SalaryOverTime) + slTinh.Rewards);
+            string Tong = tong.ToString("#,###", cul.NumberFormat);
+            MessageBox.Show("Tiền lương thực nhận của " + dgvTimeKeeping.Rows[row].Cells[1].Value.ToString() + " là: " + Tong + " VND", "", MessageBoxButtons.OK, MessageBoxIcon.Information);
         }
     }
 }
